Major refactor
- put transformers into instantiated services
- autowire transformer services to mover service
